tp: Delete unused code (#2240)
